refactor: Chrome Storage保存系ロジックの移動
