+++ src/contentScripts/index.ts
@@ -1,7 +1,7 @@
 import { setTargetBlank, startPreviewBoxObserve } from "../features/target-blank/contentScripts"
 import { getSettingsAboutTargetBlank } from "../features/target-blank/utils/getSettingsAboutTargetBlank";
 import { setTemplateSearch } from "../features/template-search/contentScripts";
-import { getSettingsAboutTemplateSearch } from "../features/template-search/hooks/getSettingsAboutTemplateSearch";
+import { getSettingsAboutTemplateSearch } from "../features/template-search/store/getSettingsAboutTemplateSearch";
 
 (async () => {
     console.log("Running 木べら！！");
