+++ src/features/template-search/store/getFavoriteTemplateList.ts
