+++ src/features/template-search/store/getSettingsAboutTemplateSearch.ts
