+++ src/features/template-search/store/setFavoriteTemplateListToChromeStorage.ts
