+++ src/features/template-search/store/setTemplateSearchSettingsToChromeStorage.ts
